+++ weed/mq/pub_balancer/balance_action.go
@@ -1,4 +1,4 @@
-package balancer
+package pub_balancer
 
 import (
 	"context"
