+++ weed/mq/pub_balancer/balance_action_split.go
@@ -1,4 +1,4 @@
-package balancer
+package pub_balancer
 
 /*
 Sequence of operations to ensure ordering
