+++ weed/mq/pub_balancer/lookup.go
@@ -1,4 +1,4 @@
-package balancer
+package pub_balancer
 
 import (
 	"errors"
